Inline output system, removing trait abstraction
Remove OutputHandler trait and separate directive/interactive handler
structs. All output logic now lives directly in global.rs with simple
match statements on the mode.

Changes:
- Delete traits.rs, directive.rs, interactive.rs
- Inline all logic into global.rs with direct mode checks
- Canonicalize shell_integration_hint() check (remove redundant guard)
- Update pre-commit exclude from directive.rs to global.rs
- Accept snapshots for shell_escape path format change

Net reduction of 305 lines while preserving all behavior.
