+++ .pre-commit-config.yaml
@@ -70,7 +70,7 @@ repos:
         types: ["rust"]
         entry: '\\x1b\['
         # format.rs strips ANSI codes injected by wrap_ansi crate
-        exclude: '^(tests/|.*_test\.rs$|src/styling/(line|mod|format)\.rs$|src/output/directive\.rs$|src/cli\.rs$)'
+        exclude: '^(tests/|.*_test\.rs$|src/styling/(line|mod|format)\.rs$|src/output/global\.rs$|src/cli\.rs$)'
 
 ci:
   # pre-commit.ci doesn't have Rust toolchain, so skip Rust-specific hooks.
